… src/crystal/system/unix/io_uring.cr
… src/crystal/system/unix/syscall/arm-linux.cr
… src/crystal/system/unix/syscall/i386-linux.cr
… src/crystal/system/unix/syscall/x86_64-linux.cr