This is a WIP implementation of the io_uring interface into Crystal's scheduler. I'm opening this PR for early review.

This was only tested with Linux 5.12 on an x86_64 machine but should work on Linux 5.5+.

Here is a benchmark to demonstrate the current performance gains:

```cr
require "http/server"

{% if flag?(:preview_iouring) %}
  require "./stdlib_patch"
{% end %}

server = HTTP::Server.new do |context|
  context.response.content_type = "text/plain"
  context.response.print "Hello world!"
end

address = server.bind_tcp 8080
puts "Listening on http://#{address}"
server.listen
```

### Before:

```
$ wrk -t12 -c100 -d60s http://127.0.0.1:8080
Running 1m test @ http://127.0.0.1:8080
  12 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.85ms  268.64us  11.66ms   88.43%
    Req/Sec     9.51k   344.83    16.35k    74.86%
  6814608 requests in 1.00m, 656.39MB read
Requests/sec: 113544.61
Transfer/sec:     10.94MB
```

### After (with `-Dpreview_iouring`):

```
$ wrk -t12 -c100 -d60s http://127.0.0.1:8080
Running 1m test @ http://127.0.0.1:8080
  12 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   422.39us  196.79us  14.97ms   95.13%
    Req/Sec    18.96k     1.96k  156.16k    92.56%
  13583654 requests in 1.00m, 1.28GB read
Requests/sec: 226021.87
Transfer/sec:     21.77MB
```

(Both tests include #10765.)

To reproduce use this PR in combination with [**stdlib_patch.cr**](https://gist.github.com/lbguilherme/adc0ac2e43d9411679798278b849cede#file-stdlib_patch-cr).

What is missing?

- [ ] Support for i386, arm, and aarch64.
- [ ] UDP Sockets.
- [ ] Connect TCP sockets.
- [ ] Connect properly with `Crystal::Scheduler` and `Crystal::EventLoop` (with flags as per https://github.com/crystal-lang/crystal/issues/10740#issuecomment-847011203).
- [ ] Add Specs.
- [ ] Validate on multiple Kernel versions.
- [ ] Solve remaining "TODO"s currently in the code.

Hopefully, this can come in time for Crystal 1.1.0 :D

Fixes #10740. Informs #10766.