@lbguilherme 2021-06-05
Updates:

This currently exposes the same primitives as LibEvent (resume, timeout, wait_readable, and wait_writeable) through almost the same interface as the original `Crystal::EventLoop`. There are no changes specifics to any particular IO. Reads, writes, accepts, connects, etc work as they have always used to: They call into LibC, if it blocks they use the event loop to wait for readiness.

This isn't the most performant way to use io_uring (every IO is still making a system call), but at least it is fully working and is already 30% faster than libevent2 on my synthetic benchmark. Doing IO directly with io_uring can be added later for extra performance.

It adds two compile-time flags and one runtime env that are only meaningful on Linux:

- `-Dpreview_iouring` enables runtime detection of a supported Kernel. It will use io_uring is available and fallback to libevent otherwise.
- `-Dforce_iouring` will always use io_uring unconditionally. On systems that don't support it, the program will fail to start.
- `CRYSTAL_DISABLE_IO_URING=1` is only meaningful when using `-Dpreview_iouring`. It disables io_uring at runtime.

For this to work `Crystal::Event` was modified to be an abstract struct and this touches the Windows port (@straight-shoota).

I couldn't enable specs because GitHub Actions runs with Kernel 5.4, which doesn't support everything we use (particularly timeout cancelation). I can probably get it working down to this version, I'll try.

Note that the diff includes everything from #10777.