@lbguilherme 2021-06-06
Got it working fine on Linux 5.4 and CI is happy :D

I'm now marking it as ready to review.

A few notes for reviewers:
- It only works in Linux 5.4 and 5.5 with `-Dforce_iouring`. Using just `-Dpreview_iouring` won't enable it because the code used for support detection only works on 5.6+. I don't think this is really an issue.
- I didn't test it much together with `-Dpreview_mt`. As both technologies are "preview" I don't think it would be wise to use them together for now. Either way, the handling is simple: one ring per thread.
- Executables build with `-Dforce_iouring` will fail to start on Linux earlier than 5.1 and will work on Linux 5.2 and 5.3, but with all timeouts resolving instantaneously. This means `sleep` will always behave as `Fiber.yield` and any IO with a timeout will fail if it needs to be async. This flag really shouldn't be used unless the person is sure it will run on Linux 5.4+.
- I'm not really sure about the behavior after a `fork`. Manpages are vague but it seems to indicate that ongoing IO operations are canceled on the child. The current implementation recreates the ring on the child to ensure it won't interfere with the parent. It at least works for a "fork-exec" and I didn't test much besides that.
- It currently exposes the same interface as libevent and doesn't take advantage of the ability to do direct IO without system calls. It's already 30% faster than libevent now. Reaching the full potential with direct IOs will be part of future PRs.
- Due to a change on `Crystal::Event` it might cause small conflicts with the Windows effort.
- The IoUring class is quite complicated and low-level, but I tried my best to explain the logic with comments. I hope it helps.