@straight-shoota 2021-06-06
Regarding the non-detection on kernel < 5.6, we could change the runtime configuration to allow forcing either alternative:

* `CRYSTAL_EVENTLOOP=io_uring` forces io_uring
* `CRYSTAL_EVENTLOOP=libevent` forces libevent
* `CRYSTAL_EVENTLOOP=auto` (or undetined) uses detection