@vlazar 2022-05-22
I've just re-discovered this great PR. Makes me sad it's still sitting here for almost a year.

@straight-shoota Is there something that's stopping this PR from being merged from core team prospective?