@lbguilherme 2022-05-24
As a reminder, there is no public API changes and everything is behind a "preview" compile flag, so there is very little risk of breaking something in production.

Either way I believe the primary concern is about manutenability in the long run, as this is a fairly complex code to adopt. It is understandable for a change like this to be not so fast to accept.

If there is anything I can do, I'm available.