@carlhoerberg 2022-10-20
```
In /usr/share/crystal/src/crystal/system/unix/event_libevent.cr:19:11

 19 | def add(timeout : Time::Span?) : Nil
              ^------
Warning: positional parameter 'timeout' corresponds to parameter 'time_span' of the overridden method Crystal::Event#add(time_span : Time::Span | ::Nil), which has a different name and may affect named argument passing
```