@straight-shoota 2022-10-21
I would very much like to merge this. More in https://github.com/crystal-lang/crystal/issues/10740#issuecomment-1287079088

Thanks for your great work pushing it forward! I think with the event loop for windows now in place we should be able to continue with the refactoring necessary for io_uring integration.