@HertzDevil 2023-05-26
Apparently there are also "I/O rings" on Windows since 21H2: https://learn.microsoft.com/en-us/windows/win32/api/ioringapi/

They are apparently very similar to `io_uring`: https://windows-internals.com/ioring-vs-io_uring-a-comparison-of-windows-and-linux-implementations/