@paulocoghi 2023-10-10
> It should be working, but there is a bug causing the CI to seg fault and couldn't find the cause yet.

Could any Crystal maintainer take a look? This improvement is of great importance.